# sdks/openapi.yaml
openapi: "3.0.3"
info:
  title: LLMSafeSpaces API
  description: |
    LLMSafeSpaces is a Kubernetes-first platform for running AI agents securely in isolated sandboxes.
    Every sandbox runs `opencode serve` as a persistent HTTP server with a PVC-backed persistent workspace.
  version: "0.7.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /workspaces/{id}:
    get:
      tags: [workspaces]
      summary: Get a workspace
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [workspaces]
      summary: Rename a workspace
      operationId: renameWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "204":
          description: Workspace renamed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [workspaces]
      summary: Delete a workspace
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "204":
          description: Workspace deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /workspaces/{id}/sessions/{sessionId}:
    get:
      tags: [sessions]
      summary: Get session details
      operationId: getSession
      x-opencode-proxy: true
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session details from opencode
          content:
            application/json:
              schema:
                type: object
                description: Opencode session object (schema tracks upstream)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [sessions]
      summary: Delete a session and its children
      description: |
        Deletes a session by proxying DELETE /session/:id to the workspace's
        opencode agent. The session and all child sessions are removed from
        both opencode and the session_index. After a successful delete, an SSE
        `session.status` event with `status: "deleted"` is published for
        multi-tab consistency.
      operationId: deleteSession
      x-opencode-proxy: true
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session deleted successfully
        "400":
          description: Invalid session ID (path traversal characters)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Workspace is not active (suspended or transitioning)
...
components:
  parameters:
    WorkspaceId:
      name: id
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Session ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login or /auth/register
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "lsp_ prefixed API key in Bearer format (e.g. Bearer lsp_abc123...)"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        userId:
          type: string
        runtime:
          type: string
        storageSize:
          type: string
        phase:
          type: string
          enum: [Pending, Creating, Active, Suspending, Suspended, Resuming, Terminating, Terminated, Failed]
        pvcName:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        devPreviewEnabled:
          type: boolean
          description: Whether the authenticated HTTP/WS dev-preview tunnel is enabled (Epic 66).
          default: false
        defaultModel:
          type: string
        agentNeedsRefresh:
          type: boolean
          description: Agent credential state (Epic 27a); always present (no omitempty).
        credentialsPendingSince:
          type: string
          format: date-time
          nullable: true
...